Move DummyRulesPrimitiveObjectParent to the tools module
