+++ aom/src/main/java/com/nedap/archie/aom/Archetype.java
@@ -1,8 +1,6 @@
 package com.nedap.archie.aom;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.google.common.base.Strings;
-import com.nedap.archie.adlparser.treewalkers.DummyRulesPrimitiveObjectParent;
 import com.nedap.archie.aom.primitives.CTerminologyCode;
 import com.nedap.archie.aom.terminology.ArchetypeTerm;
 import com.nedap.archie.aom.terminology.ArchetypeTerminology;
