+++ tools/src/test/java/com/nedap/archie/aom/TerminologyCodeConstraintsTest.java
@@ -2,7 +2,7 @@ package com.nedap.archie.aom;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
-import com.nedap.archie.adlparser.treewalkers.DummyRulesPrimitiveObjectParent;
+import com.nedap.archie.rules.evaluation.DummyRulesPrimitiveObjectParent;
 import com.nedap.archie.aom.primitives.CTerminologyCode;
 import com.nedap.archie.aom.terminology.ArchetypeTerminology;
 import com.nedap.archie.aom.terminology.ValueSet;
